removed new-common module
