+++ .idea/gradle.xml
@@ -21,7 +21,6 @@
             <option value="$PROJECT_DIR$/datastore" />
             <option value="$PROJECT_DIR$/exo-library" />
             <option value="$PROJECT_DIR$/ima-sdk" />
-            <option value="$PROJECT_DIR$/news-common" />
             <option value="$PROJECT_DIR$/omsdk" />
           </set>
         </option>
