+++ .idea/misc.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="DesignSurface">
+    <option name="filePathToZoomLevelMap">
+      <map>
+        <entry key="ad-engine/src/main/res/drawable/ic_share.xml" value="0.1325" />
+        <entry key="ad-engine/src/main/res/drawable/permission_btn_background_unselected.xml" value="0.1" />
+        <entry key="ad-engine/src/main/res/drawable/ripple_bg.xml" value="0.1" />
+        <entry key="ad-engine/src/main/res/layout/ad_brand_bottombar.xml" value="0.1" />
+        <entry key="ad-engine/src/main/res/layout/custom_snackbar_layout.xml" value="0.11875" />
+        <entry key="ad-engine/src/main/res/layout/layout_ima_video_ads.xml" value="0.3421875" />
+        <entry key="ad-engine/src/main/res/layout/like_emoji_popup_item.xml" value="0.34479166666666666" />
+        <entry key="ad-engine/src/main/res/layout/snackbar_bg.xml" value="0.11875" />
+        <entry key="app/src/main/res/layout/activity_main.xml" value="0.165625" />
+        <entry key="app/src/main/res/layout/app_actionbar_layout.xml" value="0.11666666666666667" />
+        <entry key="app/src/main/res/layout/layout_common_top_bar.xml" value="0.11666666666666667" />
+        <entry key="app/src/main/res/layout/view_dailyhunt_tab_toolbar.xml" value="0.11666666666666667" />
+        <entry key="app/src/main/res/layout/view_follow_tab_toolbar.xml" value="0.11666666666666667" />
+        <entry key="appview/src/main/res/layout/disclaimer_popup.xml" value="0.11666666666666667" />
+        <entry key="appview/src/main/res/layout/language_selection_see_all_viewholder.xml" value="0.11666666666666667" />
+        <entry key="appview/src/main/res/layout/layout_profile_ner_header.xml" value="0.1" />
+        <entry key="common/src/main/res/drawable/ic_share.xml" value="0.1335" />
+        <entry key="common/src/main/res/layout/view_cap_text_view.xml" value="0.34479166666666666" />
+        <entry key="dailyhunt-common/src/main/res/layout/layout_follow_block_button.xml" value="0.11666666666666667" />
+        <entry key="dailyhunt-common/src/main/res/layout/layout_profile_clear_activity.xml" value="0.1" />
+        <entry key="news-common/src/main/res/layout/like_emoji_popup_item.xml" value="0.1" />
+        <entry key="news-common/src/main/res/layout/onboarding_choose_language_list_item.xml" value="0.34479166666666666" />
+        <entry key="sso/src/main/res/layout/activity_true_caller_dialog.xml" value="0.175" />
+        <entry key="sso/src/main/res/layout/layout_enter_otp.xml" value="0.1" />
+      </map>
+    </option>
+  </component>
+  <component name="EntryPointsManager">
+    <list size="1">
+      <item index="0" class="java.lang.String" itemvalue="androidx.room.Dao" />
+    </list>
+  </component>
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
+    <output url="file://$PROJECT_DIR$/build/classes" />
+  </component>
+  <component name="ProjectType">
+    <option name="id" value="Android" />
+  </component>
+</project>
