+++ ad-engine/src/main/java/com/newshunt/adengine/other/dhutil/helper/BackgroundChangeListener.java
@@ -1,0 +1,14 @@
+/**
+ * Copyright (c) 2017 Newshunt. All rights reserved.
+ */
+
+package com.newshunt.adengine.other.dhutil.helper;
+
+/**
+ * Created by anshul on 07/06/17.
+ */
+
+public interface BackgroundChangeListener {
+
+  void onBackgroundColorChanged(int colorCode);
+}
