+++ ad-engine/src/main/java/com/newshunt/adengine/other/news/di/qualifiers/UiBus.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2016 Newshunt. All rights reserved.
+ */
+package com.newshunt.adengine.other.news.di.qualifiers;
+
+import javax.inject.Qualifier;
+
+/**
+ * @author satosh.dhanyamraju
+ */
+@Qualifier
+public @interface UiBus {
+}
