+++ ad-engine/src/main/java/com/newshunt/adengine/other/news/helper/MenuJsInterface.kt
@@ -1,0 +1,13 @@
+package com.newshunt.adengine.other.news.helper
+
+interface MenuJsInterface {
+    fun onMenuButtonClick(postJson: String,
+                          locationId: String,
+                          menuLocation: String,
+                          entityId: String)
+
+    fun canShowMenuButton(): Boolean
+    fun isStoryRead(postsJson: String): String
+
+    fun isStoryDisliked(postsJson: String): String
+}
