+++ ad-engine/src/main/java/com/newshunt/adengine/other/news/helper/RepostWebItemJsInterface.kt
@@ -1,0 +1,8 @@
+package com.newshunt.adengine.other.news.helper
+
+/**
+ * An interface class for the repost web item.
+ */
+interface RepostWebItemJsInterface {
+    fun onRepostClicked(postId: String)
+}
