+++ ad-engine/src/main/res/layout/layout_ima_video_ads.xml
@@ -37,7 +37,7 @@
 
         <variable
             name="vm"
-            type="com.newshunt.news.viewmodel.SocialInteractionViewModel" />
+            type="com.newshunt.adengine.other.news.viewmodel.SocialInteractionViewModel" />
 
         <variable
             name="appSettingsProvider"
