+++ ad-engine/src/main/res/layout/osv_bottom_bar.xml
@@ -13,7 +13,7 @@
 
         <import type="com.newshunt.common.helper.common.Constants" />
 
-        <import type="com.newshunt.news.helper.LikeEmojiBindingUtils" />
+        <import type="com.newshunt.adengine.other.news.helper.LikeEmojiBindingUtils" />
 
         <import type="com.newshunt.dhutil.helper.theme.ThemeUtils" />
 
@@ -48,7 +48,7 @@
 
         <variable
             name="vm"
-            type="com.newshunt.news.viewmodel.SocialInteractionViewModel" />
+            type="com.newshunt.adengine.other.news.viewmodel.SocialInteractionViewModel" />
 
     </data>
