--- news-common/src/main/AndroidManifest.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright (c) 2015 Newshunt. All rights reserved.
-  -->
-
-<manifest
-    package="com.newshunt.news.common"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <uses-permission android:name="android.permission.INTERNET"/>
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
-
-    <application>
-
-    </application>
-
-</manifest>
