--- news-common/src/main/java/com/newshunt/dhutil/view/view/BackgroundChangeListener.java
@@ -1,14 +1,0 @@
-/**
- * Copyright (c) 2017 Newshunt. All rights reserved.
- */
-
-package com.newshunt.dhutil.view.view;
-
-/**
- * Created by anshul on 07/06/17.
- */
-
-public interface BackgroundChangeListener {
-
-  void onBackgroundColorChanged(int colorCode);
-}
