--- news-common/src/main/java/com/newshunt/news/di/qualifiers/AdBus.java
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2016 Newshunt. All rights reserved.
- */
-package com.newshunt.news.di.qualifiers;
-
-import javax.inject.Qualifier;
-
-/**
- * @author satosh.dhanyamraju
- */
-@Qualifier
-public @interface AdBus {
-}
