--- news-common/src/main/java/com/newshunt/news/di/scopes/PerActivity.java
@@ -1,14 +1,0 @@
-/*
- * Copyright (c) 2018 Newshunt. All rights reserved.
- */
-
-package com.newshunt.news.di.scopes;
-
-import javax.inject.Scope;
-
-/**
- * @author satosh.dhanyamraju
- */
-@Scope
-public @interface PerActivity {
-}
