--- news-common/src/main/java/com/newshunt/news/di/scopes/PerFragment.java
@@ -1,15 +1,0 @@
-/*
- * Copyright (c) 2016 Newshunt. All rights reserved.
- */
-package com.newshunt.news.di.scopes;
-
-import javax.inject.Scope;
-
-/**
- * Scope asociated with fragment life cycle
- * @author satosh.dhanyamraju
- */
-// TODO(satosh.dhanyamraju): retention
-@Scope
-public @interface PerFragment {
-}
