--- news-common/src/main/java/com/newshunt/news/helper/MenuJsInterface.kt
@@ -1,13 +1,0 @@
-package com.newshunt.news.helper
-
-interface MenuJsInterface {
-    fun onMenuButtonClick(postJson: String,
-                          locationId: String,
-                          menuLocation: String,
-                          entityId: String)
-
-    fun canShowMenuButton(): Boolean
-    fun isStoryRead(postsJson: String): String
-
-    fun isStoryDisliked(postsJson: String): String
-}
