--- news-common/src/main/java/com/newshunt/news/helper/RepostWebItemJsInterface.kt
@@ -1,8 +1,0 @@
-package com.newshunt.news.helper
-
-/**
- * An interface class for the repost web item.
- */
-interface RepostWebItemJsInterface {
-    fun onRepostClicked(postId: String)
-}
