--- news-common/src/main/java/com/newshunt/news/helper/listeners/AdsSupportListener.java
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2016 Newshunt. All rights reserved.
- */
-package com.newshunt.news.helper.listeners;
-
-/**
- * Interface to enable/disable different type of ads based on activity.
- * @author neeraj.kumar
- */
-public interface AdsSupportListener {
-
-  boolean enableP0Ad();
-}
