--- news-common/src/main/res/values-w820dp/dimens.xml
@@ -1,10 +1,0 @@
-<!--
-  ~ Copyright (c) 2015 Newshunt. All rights reserved.
-  -->
-
-<resources>
-    <!-- Example customization of dimensions originally defined in res/values/dimens.xml
-         (such as screen margins) for screens with more than 820dp of available width. This
-         would include 7" and 10" devices in landscape (~960dp and ~1280dp respectively). -->
-    <dimen name="activity_horizontal_margin">64dp</dimen>
-</resources>
