--- news-common/src/main/res/values/color.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright (c) 2015 Newshunt. All rights reserved.
-  -->
-
-<resources>
-</resources>
