--- news-common/src/main/res/values/dimens.xml
@@ -1,6 +1,0 @@
-<!--
-  ~ Copyright (c) 2015 Newshunt. All rights reserved.
-  -->
-
-<resources>
-</resources>
