--- news-common/src/main/res/values/styles.xml
@@ -1,16 +1,0 @@
-<!--
-  ~ Copyright (c) 2015 Newshunt. All rights reserved.
-  -->
-
-<resources>
-
-
-    <!-- Base application theme. -->
-    <style name="AppTheme">
-        <!-- Customize your theme here. -->
-    </style>
-
-    <!-- TextView basic style -->
-
-
-</resources>
